tcp: start step 6, and widen the state channel for a control variable
Level 4 is the first level whose checks read a control variable rather than a
field of a segment. The first test of the kind works end to end:
Rfc5681InitialWindow.test reads the cwnd signal and passes.

Three framework gaps appeared at once, and two are repaired here.

The state channel handled intval_t alone and threw "Unsupported signal data type"
for anything else. cwnd, ssthresh and dupAcks are uintval_t; rto, srtt and rttvar
are times. Four overloads now feed one normaliser, and the event carries a
double, a time in seconds.

A scalar could be compared for equality only, while almost every rule of these
two documents states a bound: at most four segments, no less than one second.
isAtLeast, isAtMost and isBetween now express a bound.

The third gap is not repaired. A pattern cannot bind the first publication of a
signal, so a check of an initial value must forbid a bad value inside a time
window instead, which ties the check to the scenario times. This is the sharpest
framework request of the pass.

The pass also found that master does not pass its own element suite from a clean
build. WirelessLink.test and WirelessMultihop.test crash with "Unknown gate
during network initialization" here, and pass in inet-master, from identical
sources, identical tests and the same feature set. Applying the removed
PacketMultiplexer change makes them pass; reverting it makes them crash again.
The earlier report that no test depends on that change was drawn from the
inet-master build and is wrong. This pass restores nothing, because a standards
pass changes no source file.

+++ tests/protocol/lib/PacketEvent.h
@@ -54,7 +54,7 @@ struct PacketEvent {
     std::string signalName;           // the registered name of the emitting signal
     std::string sourcePath;           // emitting module path, relative to the network
     bool hasValue = false;            // true for a scalar (non-packet) signal, e.g. an FSM state
-    long value = 0;                   // the scalar signal value when hasValue
+    double value = 0;                 // the scalar signal value when hasValue; a time is in seconds
     Layer layer = Layer::Unknown;
     simtime_t time = 0;
     long treeId = -1;                 // packet identity, for cross-node correlation
